chore: added @nuxt/kit in peerDependencies
### Bugreport
For the @kyvg/vue3-notification package to work correctly, it is necessary to use one instance of nuxt both in the application and in the package. Therefore, the @nuxt/kit dependency has been moved to a peer dependency, because the project is assumed to already contain nuxt, which internally contains @nuxt/kit.